planning: folder layout change for stop-sign scenario
