+++ modules/planning/scenarios/BUILD
@@ -51,7 +51,7 @@ cc_library(
         "//modules/planning/common:planning_context",
         "//modules/planning/scenarios/lane_follow",
         "//modules/planning/scenarios/side_pass",
-        "//modules/planning/scenarios/stop_sign/stop_sign_unprotected:stop_sign_unprotected_scenario",
+        "//modules/planning/scenarios/stop_sign/unprotected:stop_sign_unprotected_scenario",
         "//modules/planning/scenarios/traffic_light/unprotected_right_turn:traffic_light_unprotected_right_turn_scenario",
     ],
 )
