+++ modules/planning/scenarios/stop_sign/unprotected/stage_pre_stop.h
@@ -20,10 +20,14 @@
 
 #pragma once
 
+#include <string>
+#include <unordered_map>
+#include <vector>
+
 #include "modules/planning/proto/planning_config.pb.h"
 
 #include "modules/planning/scenarios/stage.h"
-#include "modules/planning/scenarios/stop_sign/stop_sign_unprotected/stop_sign_unprotected_scenario.h"
+#include "modules/planning/scenarios/stop_sign/unprotected/stop_sign_unprotected_scenario.h"
 
 namespace apollo {
 namespace planning {
@@ -32,9 +36,9 @@ namespace stop_sign {
 
 struct StopSignUnprotectedContext;
 
-class StageIntersectionCruise : public Stage {
+class StagePreStop : public Stage {
  public:
-  explicit StageIntersectionCruise(const ScenarioConfig::StageConfig& config)
+  explicit StagePreStop(const ScenarioConfig::StageConfig& config)
       : Stage(config) {}
 
  private:
@@ -44,6 +48,11 @@ class StageIntersectionCruise : public Stage {
   StopSignUnprotectedContext* GetContext() {
     return GetContextAs<StopSignUnprotectedContext>();
   }
+
+  int AddWatchVehicle(const Obstacle& obstacle,
+                      std::unordered_map<std::string, std::vector<std::string>>*
+                          watch_vehicles);
+  bool CheckADCStop(const ReferenceLineInfo& reference_line_info);
 
  private:
   Stage::StageStatus FinishStage();
