+++ modules/planning/scenarios/stop_sign/unprotected/stop_sign_unprotected_scenario.cc
@@ -18,7 +18,7 @@
  * @file
  **/
 
-#include "modules/planning/scenarios/stop_sign/stop_sign_unprotected/stop_sign_unprotected_scenario.h"
+#include "modules/planning/scenarios/stop_sign/unprotected/stop_sign_unprotected_scenario.h"
 
 #include "modules/perception/proto/perception_obstacle.pb.h"
 #include "modules/planning/proto/planning_config.pb.h"
@@ -30,10 +30,10 @@
 #include "modules/planning/common/frame.h"
 #include "modules/planning/common/planning_context.h"
 #include "modules/planning/common/planning_gflags.h"
-#include "modules/planning/scenarios/stop_sign/stop_sign_unprotected/stage_creep.h"
-#include "modules/planning/scenarios/stop_sign/stop_sign_unprotected/stage_intersection_cruise.h"
-#include "modules/planning/scenarios/stop_sign/stop_sign_unprotected/stage_pre_stop.h"
-#include "modules/planning/scenarios/stop_sign/stop_sign_unprotected/stage_stop.h"
+#include "modules/planning/scenarios/stop_sign/unprotected/stage_creep.h"
+#include "modules/planning/scenarios/stop_sign/unprotected/stage_intersection_cruise.h"
+#include "modules/planning/scenarios/stop_sign/unprotected/stage_pre_stop.h"
+#include "modules/planning/scenarios/stop_sign/unprotected/stage_stop.h"
 
 namespace apollo {
 namespace planning {
