+++ modules/planning/scenarios/stop_sign/unprotected/stop_sign_unprotected_scenario.h
