+++ modules/planning/scenarios/stop_sign/unprotected/stop_sign_unprotected_scenario_test.cc
@@ -20,7 +20,7 @@
 
 #define protected public
 #define private public
-#include "modules/planning/scenarios/stop_sign/stop_sign_unprotected/stop_sign_unprotected_scenario.h"
+#include "modules/planning/scenarios/stop_sign/unprotected/stop_sign_unprotected_scenario.h"
 
 #include "gtest/gtest.h"
